Insert sorted integers into linked list
Write a C function insertSortedLL()that asks the user to input an integer, then inserts it into the linked list in ascending order. The function, insertSortedLL(), should not allow inserting an integer if it already exists in the current linked list. The function should return the index position where the new item was added; if the function could not complete successfully, it should return a value of -1. You can assume that the linked list is either a sorted linked list or an empty list.