Split a linked list into two
Write a C function frontBackSplitLinkedList() that splits the singly linked list into two sublists – one for the front half, and one for the back half. If the number of elements is odd, the extra element should go into the front list. The frontBackSplitLinkedList() prints two lists which contains frontList and backList.