Check if two binary trees are identical
Write a recursive C function identical() to determine whether two binary trees are structurally identical, assuming the two binary trees as tree1 and tree2. This function returns 1 if two binary trees are structurally identical; otherwise, it returns 0. Note that two binary trees are structurally identical if they are both empty or if they are both non-empty and the left and the right subtrees are similar (they are made of nodes with the same values and arranged in the same way).