Find the maximum height of a binary tree
Write a C function maxHeight() that accepts a pointer to the root node of a binary tree and return the number of links along the longest path from the root node down to the farthest leaf node. Note that the height of a given node is equal to the number of links from that node to the deepest leaf node. (Hint: Consider that the height of an empty tree is -1).