Create inOrderTraversal
Write an iterative C function inOrderTraversal() that prints the in-order traversal of a binary search tree using a stack. Note that you should only use push() or pop() operations when you add or remove integers from the stack. Remember to empty the stack at the beginning, if the stack is not empty.