Post-order traversal of binary tree using a stack
Write an iterative C function postOrderIterativeS1() that prints the post-order traversal of a binary search tree using a stack. Note that you should only use push() or pop() operations when you add or remove integers from the stack. Remember to empty the stack at the beginning, if the stack is not empty.