Post-order traversal of binary search tree (II)
Write an iterative C function postOrderIterativeS2() that prints the post-order traversal of a binary search tree using two stacks. Note that you should only use push() or pop() operations when you add or remove integers from the stacks. Remember to empty the stacks at the beginning, if the stacks are not empty.